feat: allow loading language-* module as plugins in application
There are many language modules LanguageTool project provided.
It is benefitical that langaugetool allow language-* module to be loaded dynimically before calling languagetool core functions.

Currently language list is built in static context.
This is proposal that the creation of languages list is happended when requested.
This allows application to load language-* modules after its launch.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>